Created Skeletons for what Client and Server Look like
There is a meeting with the team leads, I hope to get
more concrete details on what the api looks like for
auth/profile/leaderboard/matchmaking so I can build
a proper sequence diagram for each method in Server.
The Client class is really just a box for making
request calls and returning the info back.