Created Main Interface for Client and Removed Server Idea
Since Steve has said for this project we will not
need to implement a server structure for sending
message we as a group have decided to just implement
a client side interface for the networking. So we
removed the Server and made a client interface
and base implementation. It currently doesn't compile
due to not having the interfaces that the Client
has as base methods but it will be merged when we
start coding.